add the development event interceptors
Use the debug interceptor as well as an interceptor to check the app database
state against the `::db/state` spec after every event dispatch. This commit also
sets `goog.DEBUG` to false for production js builds and a `debug?` function in a
config ns to detect the dev environment and tell when the debug interceptors
should run.